renamed currentuser to currentplayer
