+++ cambio-client/src/components/CurrentPlayer/index.js
@@ -1,0 +1,5 @@
+import CurrentPlayer from './CurrentPlayer.main'
+
+export {
+  CurrentPlayer
+}
