--- cambio-client/src/components/CurrentUser/index.js
@@ -1,5 +1,0 @@
-import CurrentUser from './CurrentUser.main'
-
-export {
-  CurrentUser
-}
