+++ cambio-client/src/components/Decks/styles.scss
@@ -4,4 +4,5 @@
   border-color: $standard-deck-area;
   display: grid;
   grid-template-columns: auto 10% 10% 10% auto;
+  grid-template-rows: 100%;
 }
